Disable the Style/SingleLineBlockParams cop
This is quite an opinionated cop that on Rubocop 0.44.1 wants you to
name your reduce block params `|a, e|` and in Rubocop 0.45.0 it wants
you to name them `|acc, elem|`. So I was getting one failure locally and
a different one on Travis. Disabling this for now, then we can just use
names that make sense.

https://github.com/bbatsov/rubocop/blob/6907b13320493b334dc60a1a7687fc381f1f3740/config/default.yml#L886-L893