Rename the library to Scraped
This should avoid some of the confusion that arose from the library
being called “ScrapedPage”. It was misleading having “page” in the name
because it gave unnecessary weight to the individual pages, when
actually the tools are for scraping whole sites.

This new name means we can have classes like `Scraped::HTML`, which
make more sense for both whole pages and fragments/sections of the page.

It also has the bonus of being able to say “have you Scraped that site”
:)