Configure response decorators with a hash instead of array
Previously we wanted users to configure decorators with a two-element
array:

    [Archive, repo: 'foo/bar', branch: 'baz']

This was very simple to implement internally, but a bit opaque to the
end user looking at the code, especially the weird hash-in-an-array
syntax.

To fix this problem we're switching to instead taking an array for
configuration:

    { decorator: Archive, repo: 'foo/bar', branch: 'baz' }

This has made the _internal_ implementation (slightly) more complicated,
but externally it should be a bit clearer how this should be used.

Ideally all of this interface will be hidden behind some nicer syntax,
but for now we're sticking to the philosophy of explicit is better than
implicit, at least while we're still in the early phases of development.