Don't mutate the decorator config in ResponseDecorator
If we mutate the decorator config then it is incomplete the next time we
want to use it. Instead just pull the class out of the config hash but
don't delete it. This means that there's a useless `:decorator` key in
the decorator's config that points to `self.class`, but I don't think
that matters.