AbsoluteUrls: Handle urls that can't be parsed
This rescues any URI parsing errors and just returns the original url
rather than blowing up.

I did consider adding a warning if we're unable to parse the url, but
this made the tests very noisy and would probably be too much output on
Morph, so I've left it for now. That can be a future enhancement.