Add String#tidy method
This method is used in lots of EveryPolitician scrapers, so having it
come out of the box with scraped will help reduce duplication.

I initially attempted to do this using refinements, but couldn't get it
working in a satisfactory way. I think it's probably going to be less
confusing and require less debugging time to just do it the "old
fashioned" way and monkey patch the class directly.