[refactor] Add AbsoluteUrl class
This refactoring introduces a new `AbsoluteUrl` class. This encapsulates
the steps for joining a base url and a relative url.

This should make it easier to extract method for checking that the
relative url is a valid url according to `URI.parse`, and also helps
control the complexity of the `AbsoluteUrls#absolute_url` method.