rewrite README (includes JSON support and request headers)
Rewritten README was lagging behind changes so this is the rewrite as a single
commit. (Includes changes that had also been refelected in commits to README
for Scraped::JSON and request header passing example).

This doesn't include a separate example.rb file for tests, and it drops the
illustration which had non-exemplar example code snippets in it.