Gracefully handle broken URLs
If we try to clean up a URL that is so badly broken that URI blows up
(e.g. if it is percent-encoded as something other than UTF-8), we should
simply leave the original URL as-is, and move on, rather than generating
a run-time error on something that the user may not even care about.

See https://github.com/everypolitician/scraped/issues/99

It would also be good to log these problems in a way that is accessible
to the scraper if required, but without producing lots of unwanted
output otherwise. https://github.com/everypolitician/scraped/pull/54 is
currently dealing with that.